refac: route rename
