+++ test-web/src/routes/Stats.js
@@ -1,13 +1,13 @@
 import StatsTabs from '../components/Stats/StatsTabs';
 import Container from '@mui/material/Container';
 import Box from '@mui/material/Box';
-import SearchFilterBar from '../components/Utils/SearchFilterBar';
+import SearchFilterBar from '../components/Search/SearchFilterBar';
 import { useEffect, useState } from 'react';
 import { useLocation } from 'react-router-dom';
 import { TIME_ZONE } from '../config';
 import { Typography } from '@mui/material';
 
-function Stats() {
+const Stats = () => {
   const location = useLocation();
   const searchParams = new URLSearchParams(location.search);
 
@@ -102,6 +102,6 @@ function Stats() {
       </Box>
     </Container>
   );
-}
+};
 
 export default Stats;
